util/rlp: move ethereumjserror to rlp
