+++ packages/rlp/src/index.ts
@@ -1,3 +1,5 @@
+export * from './errors.ts'
+
 export type Input = string | number | bigint | Uint8Array | Array<Input> | null | undefined
 
 export type NestedUint8Array = Array<Uint8Array | NestedUint8Array>
